previous should call previous internally too, null check in view

+++ org.knime.core/src/eclipse/org/knime/core/node/wizard/WizardNodeView.java
@@ -370,7 +370,7 @@ public final class WizardNodeView<T extends NodeModel & WizardNode<REP, VAL>, RE
      */
     @Override
     public final void callCloseView() {
-        if (m_tempFolder.exists()) {
+        if (m_tempFolder != null && m_tempFolder.exists()) {
             deleteTempFolder(m_tempFolder);
         }
     }
